Fixed part of calculation issue
Now we are saving in the hash map diskPostings instead of only the docId because we need the term frequency for calculating the L_d of okapi algorithm.
The results are still not correct but they are more equal. We need to find the other problems. I just let in all my user outputs, maybe they help you finding some issues.