Add another check before creating the retrivalGUI
When the user wants to process a query he might be choosing a path which is not indexed yet.
Therefor we check if the the folder Indexes exists. If yes we just continue as we used to,
otherwise we let the user know that he either has to index this path first or that he has to choose another one.