Modified spelling correction
When we got the modified query of a query that contains only one corretion, it just returns the correction without the part of the query that was right.
Fixed this by looping through all terms and not only the one which need to be modified.
When I call the getModifiedQuery I will never get a null back because it is appended to a string -> check if the modified query is the same as the user input.